refactor: postTask 타입 위치 개선
