+++ src/api/axios/task-list/_types/type.ts
@@ -1,4 +1,4 @@
-import { TaskGroupItem } from "@/types";
+import { Frequency, TaskGroupItem } from "@/types";
 
 export interface GetTaskListRequest {
   groupId: string;
@@ -7,3 +7,24 @@ export interface GetTaskListRequest {
 }
 
 export type GetTaskListResponse = TaskGroupItem[];
+
+export interface PostTaskRequest {
+  groupId: string;
+  taskListId: string;
+  formData: {
+    name: string;
+    description: string;
+    startDate: string;
+    frequencyType: Frequency;
+    // TODO(지권): monthDay 테스트 필요
+    // monthDay: number;
+  };
+}
+
+export interface PostTaskResponse {
+  name: string;
+  description: string;
+  startDate: string;
+  frequencyType: Frequency;
+  monthDay: number;
+}
