+++ src/api/axios/task-list/postTask.ts
@@ -1,8 +1,8 @@
 import { instance } from "@/lib";
-import { PostTaskRequest, PostTaskResponse } from "@/types";
+import { PostTaskRequest, PostTaskResponse } from "./_types/type";
 
 const postTask = async ({ groupId, taskListId, formData }: PostTaskRequest): Promise<PostTaskResponse> => {
-  const response = await instance.post(`/groups/${groupId}/task-lists/${taskListId}/tasks`, formData);
+  const response = await instance.post<PostTaskResponse>(`/groups/${groupId}/task-lists/${taskListId}/tasks`, formData);
 
   return response.data;
 };
