+++ src/types/index.ts
@@ -10,7 +10,6 @@ export type { TaskUser } from "./TaskUser";
 export type { TaskGroupItem, TaskListData } from "./taskList/TaskListData";
 export type { DetailDataItem } from "./DetailPage/DetailDataType";
 export type { ToastVariant, AppToastOptions } from "./ToastType";
-export type { PostTaskRequest, PostTaskResponse } from "./taskList/PostTask";
 export type { TaskList } from "./Group/GroupData";
 export type { MyHistoryItem, MyHistoryResponse } from "./MyHistory/HistoryType";
 export type { ApiErrorResponse } from "./ApiType";
